+++ collect_app/src/test/java/org/odk/collect/android/formentry/InMemoryFormSessionRepositoryTest.kt
@@ -2,6 +2,7 @@ package org.odk.collect.android.formentry
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import org.junit.Rule
+import org.odk.collect.android.formentry.support.InMemFormSessionRepository
 
 class InMemoryFormSessionRepositoryTest : FormSessionRepositoryTest() {
